requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0

# Database and SQL
# Note: sqlite3 is built-in with Python, no need to install
sqlalchemy>=2.0.0

# Data processing and validation
pydantic>=2.0.0
typing-extensions>=4.0.0

# Configuration and environment
python-dotenv>=1.0.0

# Optional dependencies (install separately if needed):
# Database drivers:
# psycopg2-binary>=2.9.0  # For PostgreSQL support
# pymysql>=1.0.0  # For MySQL support

# Development tools:
# pytest>=7.0.0
# pytest-asyncio>=0.20.0
# black>=23.0.0
# flake8>=6.0.0

# Enhanced features:
# colorlog>=6.0.0  # For colored logging
# jupyter>=1.0.0  # For notebook examples
# streamlit>=1.28.0  # For web interface
# fastapi>=0.100.0  # For API server
# uvicorn>=0.20.0  # For running FastAPI